Moved JWT tests to a more appropriate place.
